Separating AppEvent from AppEventFlow.
Minor refactor.
