+++ core/cloudflow-operator-actions/src/main/scala/cloudflow/operator/event/AppEvent.scala
@@ -17,11 +17,10 @@
 package cloudflow.operator
 package event
 
-import akka.stream._
-import akka.stream.scaladsl._
+import scala.collection.immutable.Seq
+
 import skuber._
 import skuber.api.client._
-
 import cloudflow.operator.action._
 
 /**
@@ -57,51 +56,13 @@ trait AppChangeEvent[T <: ObjectResource] {
 }
 
 object AppEvent {
-
-  /**
-   * Transforms [[skuber.api.client.WatchEvent]]s into [[AppEvent]]s.
-   */
-  def fromWatchEvent(logAttributes: Attributes) =
-    Flow[WatchEvent[CloudflowApplication.CR]]
-      .statefulMapConcat { () ⇒
-        // TODO make this available in other streams directly through an Actor
-        // to prevent possibility of getting an out of sync CR from the cluster in mapToAppInSameNamespace (which might be a small chance).
-        var currentApps = Map[String, WatchEvent[CloudflowApplication.CR]]()
-        watchEvent ⇒ {
-          val cr         = watchEvent._object
-          val namespace  = cr.metadata.namespace
-          val appId      = cr.spec.appId
-          val currentApp = currentApps.get(appId).map(_._object)
-          watchEvent._type match {
-            case EventType.DELETED ⇒
-              currentApps = currentApps - appId
-              List(UndeployEvent(cr, namespace, watchEvent._object))
-            case EventType.ADDED | EventType.MODIFIED ⇒
-              if (currentApps.get(appId).forall { existingEvent ⇒
-                    existingEvent._object.resourceVersion != watchEvent._object.resourceVersion &&
-                    // the spec must change, otherwise it is not a deploy event (but likely a status update).
-                    existingEvent._object.spec != watchEvent._object.spec
-                  }) {
-                currentApps = currentApps + (appId -> watchEvent)
-                List(DeployEvent(cr, currentApp, namespace, watchEvent._object))
-              } else List()
-          }
-        }
-      }
-      .log("app-event", AppEvent.detected)
-      .withAttributes(logAttributes)
-
-  /**
-   * Transforms [[AppEvent]]s into [[Action]]s.
-   */
-  def toAction(implicit ctx: DeploymentContext): Flow[AppEvent, Action[ObjectResource], _] =
-    Flow[AppEvent]
-      .mapConcat {
-        case DeployEvent(app, currentApp, namespace, cause) ⇒
-          Actions.deploy(app, currentApp, namespace, cause)
-        case UndeployEvent(app, namespace, cause) ⇒
-          Actions.undeploy(app, namespace, cause)
-      }
+  def toActionList(appEvent: AppEvent)(implicit ctx: DeploymentContext): Seq[Action[ObjectResource]] =
+    appEvent match {
+      case DeployEvent(app, currentApp, namespace, cause) ⇒
+        Actions.deploy(app, currentApp, namespace, cause)
+      case UndeployEvent(app, namespace, cause) ⇒
+        Actions.undeploy(app, namespace, cause)
+    }
 
   def detected(appEvent: AppEvent) = s"Detected $appEvent"
 }
